improvised code
